added navbar
