--- README.md
@@ -1,12 +1,0 @@
-# BDPA Problem Statement 2021
-----------
-
-To clone repository:
-1. Console `git clone https://github.com/C4theBomb/bdpa-problem-2021.git`.
-2. Console `npm i` in both the client and backend directories.
-
-Client contains the frontend react app, backend contains all api and express code.
-
-To run project:
-1. Run `npm start` in the react directory.
-2. Run `npm run dev` in the backend directory.
