+++ client/.gitignore
@@ -1,6 +1,9 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
+**/*/node_modules
+backend/node_modules
+client/node_modules
 /node_modules
 /.pnp
 .pnp.js
@@ -17,6 +20,7 @@
 .env.development.local
 .env.test.local
 .env.production.local
+.env
 
 npm-debug.log*
 yarn-debug.log*
