+++ client/src/index.js
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom';
 import App from './App';
 
 ReactDOM.render(
-    <React.StrictMode>
-        <App />
-    </React.StrictMode>,
-    document.getElementById('root')
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  document.getElementById('root'),
 );
