--- package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "@material-ui/core": "^4.12.2",
-    "react-router-dom": "^5.2.0"
-  }
-}
